Weather-Activity/requirements.txt
FLask
requests
python-dotenv